Update constants.sh
substitute var for opt.

The idea of backing up var was for consulting logs in case of a disk failure or data wipe for instance. But since those scenarios happen without warning, it is highly unlikely that a daily backup will catch anything useful. In case of a bricked system, better to try to recover files and then, if decrypting works, check logs.